LayoutTests/imported/w3c/web-platform-tests/css/cssom/getComputedStyle-pseudo-expected.txt

Item

PASS Resolution of width is correct for ::before and ::after pseudo-elements
FAIL Pseudo-elements can use the full range of CSS syntax assert_equals: expected "50px" but got "100px"
PASS Resolution of width is correct for ::before and ::after pseudo-elements of display: contents elements
PASS Resolution of nonexistent pseudo-element styles
PASS Resolution of pseudo-element styles in display: none elements
PASS Item-based blockification of pseudo-elements
FAIL Item-based blockification of nonexistent pseudo-elements assert_equals: Pseudo-styles of display: flex elements should get blockified expected "block" but got "inline"
PASS display: contents on pseudo-elements
PASS Dynamically change to display: contents on pseudo-elements
FAIL Unknown pseudo-elements assert_true: Should return an empty style for unknown pseudo-elements starting with double-colon expected true got false
FAIL Unknown pseudo-element with a known identifier: backdrop assert_true: Should return an empty style for :backdrop expected true got false
FAIL Unknown pseudo-element with a known identifier: file-selector-button assert_true: Should return an empty style for :file-selector-button expected true got false
FAIL Unknown pseudo-element with a known identifier: grammar-error assert_true: Should return an empty style for :grammar-error expected true got false
FAIL Unknown pseudo-element with a known identifier: highlight(name) assert_true: Should return an empty style for :highlight(name) expected true got false
FAIL Unknown pseudo-element with a known identifier: marker assert_true: Should return an empty style for :marker expected true got false
FAIL Unknown pseudo-element with a known identifier: placeholder assert_true: Should return an empty style for :placeholder expected true got false
FAIL Unknown pseudo-element with a known identifier: spelling-error assert_true: Should return an empty style for :spelling-error expected true got false
FAIL Unknown pseudo-element with a known identifier: view-transition assert_true: Should return an empty style for :view-transition expected true got false
FAIL Unknown pseudo-element with a known identifier: view-transition-image-pair(name) assert_true: Should return an empty style for :view-transition-image-pair(name) expected true got false
FAIL Unknown pseudo-element with a known identifier: view-transition-group(name) assert_true: Should return an empty style for :view-transition-group(name) expected true got false
FAIL Unknown pseudo-element with a known identifier: view-transition-old(name) assert_true: Should return an empty style for :view-transition-old(name) expected true got false
FAIL Unknown pseudo-element with a known identifier: view-transition-new(name) assert_true: Should return an empty style for :view-transition-new(name) expected true got false
